Support for memory-backed file description

+++ src/libstapsdt.c
@@ -1,8 +1,9 @@
-#include <stdarg.h>
 #include <dlfcn.h>
 #include <err.h>
 #include <fcntl.h>
 #include <libelf.h>
+#include <limits.h>
+#include <stdarg.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -17,6 +18,19 @@
 #include "util.h"
 #include "libstapsdt.h"
 #include "errors.h"
+
+#ifdef LIBSTAPSDT_MEMORY_BACKED_FD
+#include <linux/memfd.h>
+#include <sys/mman.h>
+#include <sys/syscall.h>
+
+#define F_SEAL_SEAL 0x0001 /* prevent further seals from being set */
+
+// Note that linux must be 3.17 or greater to support this
+static inline int memfd_create(const char *name, unsigned int flags) {
+  return syscall(__NR_memfd_create, name, flags);
+}
+#endif
 
 int createSharedLibrary(int fd, SDTProvider_t *provider) {
   DynElf *dynElf = dynElfInit(fd);
@@ -42,6 +56,7 @@ SDTProvider_t *providerInit(const char *name) {
   provider->probes    = NULL;
   provider->_handle   = NULL;
   provider->_filename = NULL;
+  provider->_memfd = -1;
 
   provider->name = (char *) calloc(sizeof(char), strlen(name) + 1);
   memcpy(provider->name, name, sizeof(char) * strlen(name) + 1);
@@ -80,30 +95,59 @@ SDTProbe_t *providerAddProbe(SDTProvider_t *provider, const char *name, int argC
   return &(probeList->probe);
 }
 
+static char *tempElfPath(int *fd, const char *name) {
+  char *filename = NULL;
+
+#ifdef LIBSTAPSDT_MEMORY_BACKED_FD
+  char path_buffer[PATH_MAX + 1];
+  snprintf(path_buffer, (PATH_MAX + 1), "libstapsdt:%s", name);
+
+  *fd = memfd_create(path_buffer, F_SEAL_SEAL);
+  if (*fd < 0)
+    return NULL;
+  snprintf(path_buffer, (PATH_MAX + 1), "/proc/%d/fd/%d", getpid(), *fd);
+
+  filename = calloc(sizeof(char), (strlen(path_buffer) + 1));
+  strcpy(filename, path_buffer);
+#else
+  filename = calloc(sizeof(char), strlen("/tmp/-XXXXXX.so") + strlen(name) + 1);
+
+  sprintf(filename, "/tmp/%s-XXXXXX.so", name);
+
+  if ((*fd = mkstemps(filename, 3)) < 0) {
+    free(filename);
+    return NULL;
+  }
+#endif
+  return filename;
+}
+
 int providerLoad(SDTProvider_t *provider) {
   int fd;
   void *fireProbe;
-  char *filename = calloc(sizeof(char), strlen("/tmp/-XXXXXX.so") + strlen(provider->name) + 1);
   char *error;
 
-  sprintf(filename, "/tmp/%s-XXXXXX.so", provider->name);
-
-  if ((fd = mkstemps(filename, 3)) < 0) {
-    sdtSetError(provider, tmpCreationError, filename);
-    free(filename);
+  provider->_filename = tempElfPath(&fd, provider->name);
+#ifdef LIBSTAPSDT_MEMORY_BACKED_FD
+  provider->_memfd = fd;
+#endif
+  if (provider->_filename == NULL) {
+    sdtSetError(provider, tmpCreationError);
     return -1;
   }
-  provider->_filename = filename;
 
   if(createSharedLibrary(fd, provider) != 0) {
     (void)close(fd);
     return -1;
   }
-  (void)close(fd);
 
-  provider->_handle = dlopen(filename, RTLD_LAZY);
+  if (provider->_memfd == -1)
+    (void)close(fd);
+
+  provider->_handle = dlopen(provider->_filename, RTLD_LAZY);
   if (!provider->_handle) {
-    sdtSetError(provider, sharedLibraryOpenError, filename, dlerror());
+    sdtSetError(provider, sharedLibraryOpenError, provider->_filename,
+                dlerror());
     return -1;
   }
 
@@ -112,7 +156,8 @@ int providerLoad(SDTProvider_t *provider) {
 
     // TODO (mmarchini) handle errors better when a symbol fails to load
     if ((error = dlerror()) != NULL) {
-      sdtSetError(provider, sharedLibraryOpenError, filename, node->probe.name, error);
+      sdtSetError(provider, sharedLibraryOpenError, provider->_filename,
+                  node->probe.name, error);
       return -1;
     }
 
@@ -137,7 +182,12 @@ int providerUnload(SDTProvider_t *provider) {
     node->probe._fire = NULL;
   }
 
-  unlink(provider->_filename);
+  if (provider->_memfd > 0) {
+    (void)close(provider->_memfd);
+    provider->_memfd = -1;
+  } else {
+    unlink(provider->_filename);
+  }
   free(provider->_filename);
 
   return 0;
